Make the form a popover
