+++ src/components/transactions/TxDetails/Summary/index.tsx
@@ -8,7 +8,7 @@ import { Operation } from '@safe-global/safe-gateway-typescript-sdk'
 import { dateString } from '@/utils/formatters'
 import css from './styles.module.css'
 import type { SafeTransaction } from '@safe-global/safe-core-sdk-types'
-import SafeTxGasForm from '@/components/tx-flow/common/SafeTxGasForm'
+import SafeTxGasForm from '../SafeTxGasForm'
 
 interface Props {
   txDetails: TransactionDetails
@@ -76,12 +76,9 @@ export const PartialSummary = ({ safeTx }: { safeTx: SafeTransaction }) => {
   const txData = safeTx.data
   return (
     <>
-      <TxDataRow title="Operation:">{`${txData.operation} (${Operation[txData.operation].toLowerCase()})`}</TxDataRow>
       <TxDataRow title="safeTxGas:">
         <SafeTxGasForm />
       </TxDataRow>
-      <TxDataRow title="baseGas:">{txData.baseGas}</TxDataRow>
-      <TxDataRow title="refundReceiver:">{generateDataRowValue(txData.refundReceiver, 'hash', true)}</TxDataRow>
       <TxDataRow title="Raw data:">{generateDataRowValue(txData.data, 'rawData')}</TxDataRow>
     </>
   )
